requirements.txt
altair==5.0.1
asttokens==2.2.1
beautifulsoup4==4.12.2
faiss-cpu==1.7.4
langchain==0.0.212
langchainplus-sdk==0.0.17
markdown-it-py==3.0.0
streamlit==1.23.1
pypdf2==3.0.1
python-dotenv==1.0.0
pandas==2.0.2
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
prompt-toolkit==3.0.38
tiktoken==0.4.0
tqdm
streamlit-chat==0.1.1
llama-index==0.10.39
llama-parse==0.4.3